requirements.txt
fastapi
uvicorn[standard]
scikit-learn
joblib
pandas
deep-translator
python-multipart
python-dotenv
aiofiles
reportlab
pytesseract
Pillow
sqlalchemy
